zsh alias: make -> fully utilize cpu & oneline cleanup

+++ home/.zsh/alias.zsh
@@ -2,35 +2,23 @@
 alias df="df -h"
 alias ping="ping -c3"
 
-if [[ -x `which htop` ]]; then
- alias top="htop"
-fi
-
 #alias cgrep="grep --color=auto"
 alias :q="exit"
 alias :wq="exit"
 
-if [[ $(uname) = 'Darwin' ]]; then
- alias vi="mvim"
- if [[ -x `which brew` ]]; then
-  alias updateos="brew update && brew upgrade && brew cleanup"
- fi
-else
-  if [[ -x `which vim` ]]; then
-    alias vi="vim"
-  else
-    if [[ -x `which nvim` ]]; then
-      alias vi="nvim"
-    fi
-  fi
+if [[ -x `which htop` ]]; then alias top="htop" ;fi
 
-  if [[ $(uname) = 'Linux' ]]; then
-    if [[ -x `which apt` ]]; then
-      alias updateos="sudo apt update && sudo apt upgrade && sudo apt autoremove"
-    fi
-  elif [[ $(uname) = 'FreeBSD' ]]; then
-    alias updateos="echo 'freebsd not supported'"
-  fi
+if [[ $(uname) = 'Darwin' ]]; then
+  alias vi="mvim"
+  if [[ -x `which brew` ]]; then alias updateos="brew update && brew upgrade && brew cleanup" ;fi
+elif [[ $(uname) = 'Linux' ]]; then
+  NB_CORES=$(grep -c '^processor' /proc/cpuinfo)
+  alias make="make -j$((NB_CORES+1)) -l${NB_CORES}"
+  if [[ -x `which apt` ]]; then alias updateos="sudo apt update && sudo apt upgrade && sudo apt autoremove" ;fi
+  if [[ -x `which vim` ]]; then alias vi="vim" ; elif [[ -x `which nvim` ]]; then alias vi="nvim" ;fi
+elif [[ $(uname) = 'FreeBSD' ]]; then
+  alias updateos="echo 'freebsd not supported'"
+  if [[ -x `which vim` ]]; then alias vi="vim" ; elif [[ -x `which nvim` ]]; then alias vi="nvim" ;fi
 fi
 
 if type -p colordiff &> /dev/null ; then
@@ -40,12 +28,11 @@ else
     alias diff="vimdiff" ;
   fi
 fi
-if type -p colorgcc &> /dev/null ; then alias gcc="colorgcc" ; fi
-if type -p colortail &> /dev/null ; then alias tail="colortail" ; fi
 
+if type -p colorgcc &> /dev/null ; then alias gcc="colorgcc" ;fi
+if type -p colortail &> /dev/null ; then alias tail="colortail" ;fi
 
 if [[ -x `which git` ]]; then
- alias gs="git status"
- alias gp="git pull && git submodule update --init"
+ alias gs="git status" ;
+ alias gp="git pull && git submodule update --init" ;
 fi
-
